requirements.txt
httpx
python-dotenv
black
isort
mypy
flake8
pytz
types-pytz
